+++ CHANGELOG.md
@@ -295,7 +295,7 @@ Upgrade library dependencies (if required):
 
 - Fixed issue when websockets are enabled making to many API calls (for those that are still required they will be called once per minute instead of on every iteration).
 - Added telegram bot (telegram_bot.py) this has to be run separately to pycryptobot.py
-- Added option to enable the telegram bot this does not start it, just enables writing data required for the telegram bot (--enable_telegram_bot_control)
+- Added option to enable the telegram bot this does not start it, just enables writing data required for the telegram bot (--telegrambotcontrol)
 - Added user_id to telegram section in config, this is used so that only you can control the bot
 - Added datafolder to telegram section in config, this is used when running multiple bots from different folders
 - Added Telegram Bot Setup instructions text file